Rename testdata folder to contracts
