+++ op-e2e/interop/contracts/emit.go
