+++ op-e2e/interop/contracts/generate.sh
