+++ op-e2e/interop/contracts/test-artifacts/emit.sol/EmitEvent.bin
